requirements.txt
flask==2.3.2
celery
pandas==2.0.3
numpy<2
flask-restx==1.1.0
pyarrow==12.0.1
scikit-learn==1.0.2
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.12
alembic==1.8.1
psycopg2-binary==2.9.3
Flask-Migrate==3.1.0
pytest==7.4.0
gunicorn==21.2.0
confluent-kafka
kombu
Werkzeug==2.3.7
marshmallow
python-dotenv==1.0.1
flask-cors